rpc/httpkit: support generic HTTP error responses
The ErrorEncoder is now able to return as JSON any error object
that implements the json.Marshaller interface.

As a helper the httpkit package now provides and HttpError object
that implements the json.Marshaller and gokit HTTP interfaces like
StatusCoder, Headerer to provide a way the client code to define it's
own error messages.